requirements.txt
python-dotenv
psycopg2-binary
tqdm
chia-blockchain